MCR-2410 move getCreationPredicate() and getRegistrationPredicate()
from MCRPIJobService to MCRPIService
